move context files to context folders
