+++ src/core/storage/disk.ts
@@ -4,8 +4,7 @@ import fs from "fs/promises"
 import { Anthropic } from "@anthropic-ai/sdk"
 import { fileExistsAtPath } from "../../utils/fs"
 import { ClineMessage } from "../../shared/ExtensionMessage"
-import type { TaskMetadata } from "../context-tracking/ContextTrackerTypes"
-
+import { TaskMetadata } from "../context/context-tracking/ContextTrackerTypes"
 export const GlobalFileNames = {
 	apiConversationHistory: "api_conversation_history.json",
 	contextHistory: "context_history.json",
